Move Pair tests into the Module tests.
